Egg Monitoring System
Idea: The system will monitor the egg environment to produce the best egg product. The DHT22 will measure the temperature and humidity of the egg/egg container. At the same time, the ultrasonic will detect if any obstacles/animals/bugs can destroy the egg. For normal conditions, the green LED will light on and the other output components will remain turned off. If there are any obstacles detected nearly 10cm from the ultrasonic, the system will trigger the triggerAlert functions where the green LED will turn off, the red LED will light on, and the servo will move 180 degrees to get rid of the obstacles/animals and the buzzer will turn on to give alert to the monitor/owner.